+++ src/main/java/io/unthrottled/theme/randomizer/ui/PluginSettingsUI.java
@@ -14,8 +14,8 @@ import io.unthrottled.theme.randomizer.config.IntervalTuple;
 import io.unthrottled.theme.randomizer.config.SettingsHelper;
 import io.unthrottled.theme.randomizer.config.actors.LafAnimationActor;
 import io.unthrottled.theme.randomizer.mode.PluginMode;
-import io.unthrottled.theme.randomizer.services.SystemMatchManager;
-import io.unthrottled.theme.randomizer.services.ThemeGatekeeper;
+import io.unthrottled.theme.randomizer.system.match.SystemMatchManager;
+import io.unthrottled.theme.randomizer.themes.ThemeGatekeeper;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
