+++ src/main/kotlin/io/unthrottled/theme/randomizer/themes/GetId.kt
@@ -1,0 +1,10 @@
+package io.unthrottled.theme.randomizer.themes
+
+import com.intellij.ide.ui.laf.UIThemeBasedLookAndFeelInfo
+import javax.swing.UIManager
+
+fun UIManager.LookAndFeelInfo.getId(): String =
+  when (this) {
+    is UIThemeBasedLookAndFeelInfo -> this.theme.id
+    else -> this.name
+  }
