+++ src/main/kotlin/io/unthrottled/theme/randomizer/themes/SelectableThemeType.kt
@@ -1,0 +1,6 @@
+package io.unthrottled.theme.randomizer.themes
+
+enum class SelectableThemeType {
+  DARK, LIGHT, ANY
+}
+
