+++ src/main/kotlin/io/unthrottled/theme/randomizer/timed/ThemeChangeEventEmitter.kt
@@ -1,4 +1,4 @@
-package io.unthrottled.theme.randomizer.services
+package io.unthrottled.theme.randomizer.timed
 
 import com.intellij.ide.IdeEventQueue
 import com.intellij.ide.actions.QuickChangeLookAndFeel
@@ -10,9 +10,10 @@ import com.intellij.util.Alarm
 import io.unthrottled.theme.randomizer.config.ChangeIntervals
 import io.unthrottled.theme.randomizer.config.Config
 import io.unthrottled.theme.randomizer.config.ConfigListener
-import io.unthrottled.theme.randomizer.config.ConfigListener.Companion.CONFIG_TOPIC
 import io.unthrottled.theme.randomizer.mode.PluginMode
 import io.unthrottled.theme.randomizer.mode.toPluginMode
+import io.unthrottled.theme.randomizer.themes.ThemeService
+import io.unthrottled.theme.randomizer.themes.getId
 import java.time.Duration
 import java.time.Instant
 import java.util.concurrent.TimeUnit
@@ -39,7 +40,7 @@ class ThemeChangeEventEmitter : Runnable, LafManagerListener, Disposable {
   init {
     val self = this
     messageBus.subscribe(
-      CONFIG_TOPIC,
+        ConfigListener.CONFIG_TOPIC,
       ConfigListener { newPluginState, _ ->
         themeChangeAlarm.cancelAllRequests()
         if (newPluginState.isChangeTheme &&
@@ -55,7 +56,7 @@ class ThemeChangeEventEmitter : Runnable, LafManagerListener, Disposable {
       }
     )
     messageBus.subscribe(
-      LafManagerListener.TOPIC,
+        LafManagerListener.TOPIC,
       this,
     )
     themeChangeAlarm.addRequest(
@@ -84,7 +85,7 @@ class ThemeChangeEventEmitter : Runnable, LafManagerListener, Disposable {
 
   private fun convertMinutesToMillis(duration: Long) = TimeUnit.MILLISECONDS.convert(
     duration,
-    TimeUnit.MINUTES
+      TimeUnit.MINUTES
   ).toInt()
 
   private fun getThemeSwitchCheckInterval(): Long =
@@ -113,9 +114,9 @@ class ThemeChangeEventEmitter : Runnable, LafManagerListener, Disposable {
     ThemeService.instance.nextTheme()
       .ifPresent {
         QuickChangeLookAndFeel.switchLafAndUpdateUI(
-          LafManager.getInstance(),
-          it,
-          true
+            LafManager.getInstance(),
+            it,
+            true
         )
         themeSet = it
         captureTimestamp()
@@ -127,8 +128,8 @@ class ThemeChangeEventEmitter : Runnable, LafManagerListener, Disposable {
     getThemeChangeIntervalInMinutes(config) <= getDurationSinceThemeChange().toMinutes()
 
   private fun getDurationSinceThemeChange() = Duration.between(
-    Instant.ofEpochSecond(Config.instance.lastChangeTime),
-    Instant.now()
+      Instant.ofEpochSecond(Config.instance.lastChangeTime),
+      Instant.now()
   )
 
   @SuppressWarnings("MagicNumber")
