+++ src/main/resources/META-INF/plugin.xml
@@ -15,8 +15,8 @@
     <applicationService serviceImplementation="io.unthrottled.theme.randomizer.config.Config"/>
     <applicationService serviceImplementation="io.unthrottled.theme.randomizer.services.LAFProbabilityService"/>
     <applicationService serviceImplementation="io.unthrottled.theme.randomizer.PluginMaster"/>
-    <applicationService serviceImplementation="io.unthrottled.theme.randomizer.services.ThemeGatekeeper"/>
-    <applicationService serviceImplementation="io.unthrottled.theme.randomizer.services.ThemeService"/>
+    <applicationService serviceImplementation="io.unthrottled.theme.randomizer.themes.ThemeGatekeeper"/>
+    <applicationService serviceImplementation="io.unthrottled.theme.randomizer.themes.ThemeService"/>
     <notificationGroup displayType="BALLOON" id="Theme Randomizer Updates" toolWindowId="Theme Randomizer Updates" icon="ThemeRandomizerIcons.PLUGIN_ICON" isLogByDefault="false" />
     <errorHandler implementation="io.unthrottled.theme.randomizer.integrations.ErrorReporter"/>
   </extensions>
